# Add stream additions