--- a/stream.go
+++ b/stream.go
@@ -5,12 +5,13 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
-	"github.com/mediocregopher/radix/v3/internal/bytesutil"
-	"github.com/mediocregopher/radix/v3/resp"
-	"github.com/mediocregopher/radix/v3/resp/resp2"
 	"io"
 	"math"
 	"strconv"
+
+	"github.com/mediocregopher/radix/v3/internal/bytesutil"
+	"github.com/mediocregopher/radix/v3/resp"
+	"github.com/mediocregopher/radix/v3/resp/resp2"
 )
 
 // StreamEntryID represents an ID used in a Redis stream with the format <time>-<seq>.