Fixes #47