@nussjustin 2018-12-08
Most things done. Currently checking the impact of `resp2.Any`. 

Impact for `StreamEntry` unmarshaling (old is current, new is with `Any`):

```
name           old time/op    new time/op    delta
StreamEntry-8    30.6µs ± 1%    32.7µs ± 1%    +6.81%  (p=0.008 n=5+5)

name           old alloc/op   new alloc/op   delta
StreamEntry-8     64.0B ± 0%    128.0B ± 0%  +100.00%  (p=0.008 n=5+5)

name           old allocs/op  new allocs/op  delta
StreamEntry-8      5.00 ± 0%      9.00 ± 0%   +80.00%  (p=0.008 n=5+5)
```