@nussjustin 2018-12-08
For StreamReader:

```
name            old time/op    new time/op    delta
StreamReader-8     119µs ± 1%     124µs ± 1%  +3.84%  (p=0.008 n=5+5)

name            old alloc/op   new alloc/op   delta
StreamReader-8    14.8kB ± 0%    14.9kB ± 0%  +0.22%  (p=0.008 n=5+5)

name            old allocs/op  new allocs/op  delta
StreamReader-8       145 ± 0%       146 ± 0%  +0.69%  (p=0.008 n=5+5)
```