@nussjustin 2018-12-08
Ok, I optimized StreamReader version with Any a bit (and cleaned it up!) so that it now allocates less than the current implementation:

```
name            old time/op    new time/op    delta
StreamReader-8     119µs ± 1%     124µs ± 1%  +3.77%  (p=0.008 n=5+5)

name            old alloc/op   new alloc/op   delta
StreamReader-8    14.8kB ± 0%    14.5kB ± 0%  -1.94%  (p=0.008 n=5+5)

name            old allocs/op  new allocs/op  delta
StreamReader-8       145 ± 0%       145 ± 0%    ~     (all equal)
```