@nussjustin 2018-12-08
Pushed the cleaned up streamReaderEntry unmarshaling incl. the `resp2.Any` use. Now it even allocates less than before:

```
name            old time/op    new time/op    delta
StreamReader-8     119µs ± 1%     123µs ± 1%  +3.37%  (p=0.008 n=5+5)

name            old alloc/op   new alloc/op   delta
StreamReader-8    14.8kB ± 0%    14.5kB ± 0%  -2.16%  (p=0.008 n=5+5)

name            old allocs/op  new allocs/op  delta
StreamReader-8       145 ± 0%       144 ± 0%  -0.69%  (p=0.008 n=5+5)
```

And this is still without one optimization that I removed (reusing allocated strings for stream names).