@nussjustin 2018-12-23
Travis still times out, but all test pass locally (with and without `-race`).

It seems like the compilation itself is the problem on Travis since the test themselves start outputting stuff immediately (this can be easily seen when splitting compilation and execution, e.g. via `go test -c -race` and `./v3.test -test.count 1 -test.v`). Maybe adding `-x` to the `go test -v -race` command could help identify the problem.