@mediocregopher 2018-12-23
I think it has something to do with the `./...` portion of the `go test -v race ./...` line.

- Running `go test -v -x ./...` gives [this](https://travis-ci.org/mediocregopher/radix/builds/471641133). It seems to get blocked removing a directory related to the resp2 test build?

- Running `go test -v -x .` gives [this](https://travis-ci.org/mediocregopher/radix/builds/471642773). The build works but the tests themselves fail because XADD isn't a command (???)

- Running just `go test -v -x ./resp/resp2` gives [this](https://travis-ci.org/mediocregopher/radix/builds/471644321), a successful build.

So it doesn't like something about `./...`, but the tests themselves seem broken as well, maybe redis isn't the right version on travis?