@nussjustin 2018-12-23
Line 360 in the build output (unter "Build system information") says

```
Redis version
redis-server 4.0.6
```

So the Redis version is to old, so we should probably the current stable manually.