@nussjustin 2018-12-23
Regarding the timeout: Did you try a something like `go list ./... | xargs -n1 go test -race -v`? This would compile and run each packages tests independently.