@mediocregopher 2018-12-24
`go list` was a good idea, I got it working:

```yml
---
language: go
go:
  - "1.11"
  - "1.10.7"

services:
  - redis-server

before_install:
  # update to latest version of redis
  - sudo apt-get install -y redis-server

script:
  - go list ./... | xargs -n1 go test -v -race

after_failure:
  - tail -n100 ./*.log
```

that also adds testing for 1.10 which is nice. You'll also need to revert 6fcfd0c, sorry =/ The timeout stuff will have to be dealt with separately, but the PR works without it.