@mediocregopher 2018-12-29
Went ahead and updated the travis file in master and merged this, even though it never built I'm confident it will once it's in master.